pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viber"
version = "0.1.0"
description = "Viber data platform scaffold"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Viber Team" }]
dependencies = [
  "azure-identity>=1.16",
  "azure-storage-file-datalake>=12.14",
  "pandas>=2.0",
  "pyarrow>=15.0",
]

[project.urls]
Homepage = "https://example.com"

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "ruff>=0.3",
  "mypy>=1.8",
  "pre-commit>=3.5",
]

[tool.ruff]
line-length = 100
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["viber"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
pythonpath = ["src"]
